Introduce scraping and summarisation tools
